git-svn-id: https://www.seasar.org/svn/sandbox/s2chronos/trunk@168 5fbd9426-1b17-45cb-9c0a-1918607842a3

+++ s2chronos-extension/src/main/java/org/seasar/chronos/extension/store/TriggerStore.java
@@ -25,7 +25,7 @@ public class TriggerStore {
 
 	private TimedTriggerDxo timedTriggerDxo;
 
-	public TaskTrigger loadFromStore(int id) {
+	public TaskTrigger loadFromStore(Long id) {
 		TriggerEntity triggerEntity = triggerDao.selectById(id);
 		if (triggerEntity == null) {
 			return null;
@@ -42,7 +42,7 @@ public class TriggerStore {
 		return null;
 	}
 
-	public void loadFromStore(int id, TaskTrigger trigger) {
+	public void loadFromStore(Long id, TaskTrigger trigger) {
 		TriggerEntity triggerEntity = triggerDao.selectById(id);
 		if ("CT".equals(triggerEntity.getExecType())) {
 			cronTriggerDxo.fromEntityFromComponent(triggerEntity,
